+++ playwright/src/util.ts
@@ -1,0 +1,5 @@
+function range(size: number, startAt: number = 0): ReadonlyArray<number> {
+  return [...Array(size).keys()].map((i) => i + startAt);
+}
+
+export { range };